+++ src/extensionConsole/pages/integrations/PrivateIntegrationsCard.tsx
@@ -24,7 +24,7 @@ import PaginatedTable from "@/components/paginatedTable/PaginatedTable";
 import { type RootState } from "@/store/optionsStore";
 import { faEdit, faSignOutAlt } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { deleteCachedAuthData } from "@/background/messenger/api";
+import { deleteCachedAuthData } from "@/background/messenger/strict/api";
 import notify from "@/utils/notify";
 import EllipsisMenu from "@/components/ellipsisMenu/EllipsisMenu";
 import BrickIcon from "@/components/BrickIcon";
