+++ src/hooks/useTheme.ts
@@ -20,7 +20,7 @@ import { selectSettings } from "@/store/settings/settingsSelectors";
 import settingsSlice from "@/store/settings/settingsSlice";
 import { useDispatch, useSelector } from "react-redux";
 import { DEFAULT_THEME, type Theme } from "@/themes/themeTypes";
-import { activatePartnerTheme } from "@/background/messenger/api";
+import { activatePartnerTheme } from "@/background/messenger/strict/api";
 import {
   addThemeClassToDocumentRoot,
   getThemeLogo,
